refactor(workspace): split the deterministic templates out of crew_roles
`crew_roles.rs` reached 1262 lines and `compliance.rs` 1214, both past the
1200-line ceiling. The repo's own rule is to split before the limit, not
after the gate says so.

`crew_template.rs` takes the deterministic drafts — the bounded proposals
the product writes when no model answers or a model's answer is refused —
along with the two unit tests that cover their helpers. The cut is there
because the dependency runs one way: the templates read a `RoleInput`,
and nothing in the role module refers to a template. `MAX_LIST_ITEMS`
becomes `pub(super)` so one cap still bounds both a model's list and a
template's. Each module keeps its own private `invalid`, as the other
four modules here already do.

`compliance.rs` keeps its shape; its unit tests move to a child module,
`compliance/tests.rs`, so they still reach `tokens`, `retrieval` and
`validate_summary` without any of those becoming visible to the crate.

A pure move: the same 68 workspace tests pass before and after. What is
left is queued — `compliance.rs` is at 1143 and its next feature would
hit the ceiling again, so the follow-up cut is written down rather than
discovered by the gate later.
